refactor(google_vertex): extract model-catalog group to google_vertex_catalog.py
Phase 0c-2 thinning (slice 1): move the model-fetch/normalize/static-config group
(fetch_models_from_google_vertex + helpers, ~371 lines, off the hot path) out of
the 1784-line client into google_vertex_catalog.py. Redirect model_catalog_sync's
import + the multi-endpoint test's import to the new module; no import cycle (the
client does not import the catalog). client 1785 -> 1411. Verified against the
full local suite (zero regressions; google_vertex tests collect + pass).
